src/main/java/com/example/servlet/domain/member/MemberRepository.java
package com.example.servlet.domain.member;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.ConcurrentMap;
import java.util.concurrent.atomic.AtomicLong;

/*
동시성 문제를 해결하기위해 HashMap, long 대신 ConcurrentHashMap, AtomicLong 사용
Singleton으로 생성
 */
public class MemberRepository {

    private ConcurrentMap<Long, Member> store = new ConcurrentHashMap<>();
    private AtomicLong sequence = new AtomicLong();

    private static final MemberRepository instance = new MemberRepository();

    private MemberRepository() {}

    public static MemberRepository getInstance() {
        return instance;
    }

    public Member save(Member member) {
        member.setId(sequence.incrementAndGet());
        store.put(member.getId(), member);
        return member;
    }

    public Member findBy(Long id) {
        return store.get(id);
    }

    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
